moved sidebar and chat components into feature files
