+++ frontend/src/chat/EmptyChatState.jsx
@@ -1,0 +1,5 @@
+export function EmptyChatState() {
+  return (
+    <div className="min-h-[100dvh]" aria-hidden="true" />
+  );
+}
