+++ frontend/src/main.jsx
@@ -1,48 +1,3 @@
-import { cx, CONTROL_MOTION, FADE_MOTION } from "./uiShared.js";
-import { useState, useRef, useEffect, useMemo, useCallback, memo } from "react";
-import {
-  formatInteger,
-  formatCost,
-  truncatePromptText,
-  formatThoughtDuration,
-  exportFileName,
-  shortTitle,
-  storyExportFileName,
-} from "./textFormatting.js";
-import { OverflowActions } from "./components/OverflowActions.jsx";
-import {
-  Pencil,
-  X,
-  Trash2,
-  Eye,
-  EyeOff,
-  PanelLeftOpen,
-  ChevronDown,
-  Copy,
-  RefreshCw,
-  Menu,
-} from "lucide-react";
-import { MaskIcon, IconButton } from "./components/IconButton.jsx";
-import { createPortal } from "react-dom";
-import { MarkdownContent } from "./chat/MarkdownContent.jsx";
-import { api, responseErrorDetail } from "./api.js";
-import {
-  promptModelName,
-  getModelContextLimit,
-  toFiniteNumber,
-  getContextWindowInfo,
-  isFreeModel,
-} from "./modelFormatting.js";
-import { APP_VERSION } from "./appInfo.js";
-import { SlidingTabs } from "./components/SlidingTabs.jsx";
-import { CHAT_MODES, newSettings, DEFAULT_MODEL } from "./settings/settingsDefaults.js";
-import { NamePromptModal } from "./components/NamePromptModal.jsx";
-import { StatusLabel } from "./components/StatusLabel.jsx";
-import ThinkingContent from "./ThinkingContent.jsx";
-import { uncitedSources } from "./websearch/citations.js";
-import { useStreamReveal } from "./streamingText.js";
-import AttachmentChips from "./attachments/AttachmentChips.jsx";
-import SourcePills from "./websearch/SourcePills.jsx";
 import {
   readLocalAppSettings,
   pickOpeningMessage,
@@ -51,6 +6,8 @@ import {
   clearLocalChatFolders,
   writeLocalAppSettings,
 } from "./localStorage.js";
+import { useState, useEffect, useRef, useMemo, useCallback } from "react";
+import { newSettings, DEFAULT_MODEL } from "./settings/settingsDefaults.js";
 import { parseRoute, storyRoute, routePath, chatRoute } from "./routing.js";
 import { useTour } from "./tour/useTour.js";
 import { WRITE_TOUR_STEPS } from "./tour/tourSteps.js";
@@ -79,15 +36,33 @@ import {
   chapterGenerationErrorIsRepairable,
   chapterAppliedEditSummary,
 } from "./writing/chapterGenerationEvents.js";
+import {
+  getModelContextLimit,
+  toFiniteNumber,
+  getContextWindowInfo,
+  isFreeModel,
+  promptModelName,
+} from "./modelFormatting.js";
+import { api, responseErrorDetail } from "./api.js";
+import { exportFileName, shortTitle, storyExportFileName } from "./textFormatting.js";
 import { updateLorebookStream } from "./lorebook/lorebookUpdateApi.js";
 import { repairLorebook as repairLorebookStream } from "./lorebook/repairLorebookApi.js";
 import { generateLorebookEntry as generateLorebookEntryStream } from "./lorebook/generateEntryApi.js";
 import { useFileDrop } from "./attachments/useFileDrop.js";
+import { StoryRail } from "./sidebar/StoryRail.jsx";
+import { ConversationRail } from "./sidebar/ConversationRail.jsx";
+import { SidebarRevealButton } from "./sidebar/SidebarRevealButton.jsx";
+import { IconButton } from "./components/IconButton.jsx";
+import { Menu } from "lucide-react";
 import HelpTourButton from "./HelpTourButton.jsx";
+import { TemporaryChatButton, TemporaryChatMarker } from "./chat/TemporaryChatButton.jsx";
 import StoryBrainstorm from "./brainstorm/StoryBrainstorm.jsx";
 import { ContextWindowMeter } from "./components/ContextWindowMeter.jsx";
 import { StoryWorkspace } from "./writing/StoryWorkspace.jsx";
 import { LOREBOOK_PHASE_LABELS } from "./writing/WriteOperationStatus.jsx";
+import { MessageList } from "./chat/MessageList.jsx";
+import { PromptNavigationRail } from "./chat/PromptNavigationRail.jsx";
+import { EmptyChatState } from "./chat/EmptyChatState.jsx";
 import { WriteLanding } from "./writing/WriteLanding.jsx";
 import { Composer } from "./composer/Composer.jsx";
 import { SettingsDrawer } from "./settings/SettingsDrawer.jsx";
@@ -98,2510 +73,6 @@ import TourOverlay from "./tour/TourOverlay.jsx";
 import { TosLoadingScreen, TosUnavailableScreen, TosGateModal } from "./TosGate.jsx";
 import { createRoot } from "react-dom/client";
 import "./styles.css";
-
-const FEEDBACK_FORM_URL = "https://forms.gle/gTth2TcXLYAArvGm6";
-
-const GITHUB_RELEASES_LATEST_URL = "https://api.github.com/repos/echo1097/routerchat/releases/latest";
-
-function AssistantActionButton({ label, children, ...props }) {
-  return (
-    <button
-      aria-label={label}
-      title={label}
-      className={cx(
-        "inline-flex h-6 w-6 items-center justify-center rounded-md text-neutral-500 hover:text-neutral-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-white/15",
-        CONTROL_MOTION,
-      )}
-      {...props}
-    >
-      {children}
-    </button>
-  );
-}
-
-function ResponseInfoButton({ message }) {
-  const [open, setOpen] = useState(false);
-  const [closing, setClosing] = useState(false);
-  const rootRef = useRef(null);
-  const closeTimerRef = useRef(null);
-  const popoverId = `response-info-${message.id}`;
-
-  const rows = [
-    ["Total input tokens", formatInteger(message.prompt_tokens)],
-    ["Total output tokens", formatInteger(message.completion_tokens)],
-    ["Total tokens", formatInteger(message.total_tokens)],
-    ["Total cost", formatCost(message.cost)],
-    ["Model", message.model || "Unavailable"],
-  ];
-
-  function clearCloseTimer() {
-    if (!closeTimerRef.current) return;
-    window.clearTimeout(closeTimerRef.current);
-    closeTimerRef.current = null;
-  }
-
-  function closePopover() {
-    if (!open) return;
-    clearCloseTimer();
-    setOpen(false);
-    setClosing(true);
-    const closeMs = parseFloat(
-      getComputedStyle(document.documentElement).getPropertyValue("--dropdown-close-dur"),
-    ) || 150;
-    closeTimerRef.current = window.setTimeout(() => {
-      setClosing(false);
-      closeTimerRef.current = null;
-    }, closeMs);
-  }
-
-  function togglePopover(event) {
-    event.stopPropagation();
-    if (open) {
-      closePopover();
-      return;
-    }
-    clearCloseTimer();
-    setClosing(false);
-    setOpen(true);
-  }
-
-  useEffect(() => {
-    if (!open) return undefined;
-
-    function handlePointerDown(event) {
-      if (rootRef.current?.contains(event.target)) return;
-      closePopover();
-    }
-
-    function handleKeyDown(event) {
-      if (event.key === "Escape") closePopover();
-    }
-
-    document.addEventListener("pointerdown", handlePointerDown);
-    document.addEventListener("keydown", handleKeyDown);
-    return () => {
-      document.removeEventListener("pointerdown", handlePointerDown);
-      document.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [open]);
-
-  useEffect(() => () => clearCloseTimer(), []);
-
-  return (
-    <div ref={rootRef} className="response-info">
-      <button
-        type="button"
-        aria-label={open ? "Close response info" : "Show response info"}
-        title={open ? "Close response info" : "Show response info"}
-        aria-expanded={open}
-        aria-controls={popoverId}
-        onClick={togglePopover}
-        className={cx(
-          "inline-flex h-6 w-6 items-center justify-center rounded-md text-neutral-500 hover:text-neutral-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-white/15",
-          CONTROL_MOTION,
-        )}
-      >
-        <span className="t-icon-swap response-info-icon" data-state={open ? "b" : "a"}>
-          <span className="t-icon" data-icon="a" aria-hidden="true">
-            <i className="fi fi-rc-info" />
-          </span>
-          <span className="t-icon" data-icon="b" aria-hidden="true">
-            <i className="fi fi-br-cross-small" />
-          </span>
-        </span>
-      </button>
-      <div
-        id={popoverId}
-        role="dialog"
-        aria-label="Response information"
-        data-origin="bottom-left"
-        className={cx(
-          "t-dropdown response-info-popover",
-          open && "is-open",
-          closing && "is-closing",
-        )}
-      >
-        <div className="response-info-title">Response info</div>
-        <dl className="response-info-grid">
-          {rows.map(([label, value]) => (
-            <div key={label} className="response-info-row">
-              <dt>{label}</dt>
-              <dd>{value}</dd>
-            </div>
-          ))}
-        </dl>
-      </div>
-    </div>
-  );
-}
-
-function ChatHistoryActions({
-  chat,
-  isFirst,
-  forceVisible,
-  folders,
-  renameDisabled,
-  onRename,
-  onDelete,
-  onExport,
-  onTogglePin,
-  onMoveToFolder,
-}) {
-  const [moveOpen, setMoveOpen] = useState(false);
-  return (
-    <OverflowActions
-      id={`chat-${chat.id}`}
-      title={chat.title}
-      label="Chat actions"
-      isFirst={isFirst}
-      forceVisible={forceVisible}
-    >
-      {(closeMenu) => (
-        <>
-          <button
-            type="button"
-            role="menuitem"
-            disabled={renameDisabled}
-            title={renameDisabled ? "Waiting for the chat name" : undefined}
-            onClick={() => {
-              closeMenu();
-              onRename(chat);
-            }}
-            className={cx(
-              "chat-history-menu-item focus:outline-none",
-              renameDisabled
-                ? "cursor-not-allowed text-neutral-600"
-                : "text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07]",
-            )}
-          >
-            <Pencil size={14} />
-            Edit name
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onTogglePin(chat);
-            }}
-            className="chat-history-menu-item text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-          >
-            <i className="fi fi-rr-thumbtack text-[14px] leading-none" aria-hidden="true" />
-            {chat.pinned ? "Unpin chat" : "Pin chat"}
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            aria-expanded={moveOpen}
-            onClick={() => setMoveOpen((current) => !current)}
-            className="chat-history-menu-item text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-          >
-            <MaskIcon src="/icons/folder.png" size={14} />
-            Move to folder
-          </button>
-          {moveOpen && (
-            <div className="border-l border-white/10 pl-2">
-              {folders.length === 0 && (
-                <div className="px-2 py-1.5 text-[12px] leading-4 text-neutral-500">
-                  No folders yet
-                </div>
-              )}
-              {folders.map((folder) => (
-                <button
-                  key={folder.id}
-                  type="button"
-                  role="menuitem"
-                  onClick={() => {
-                    setMoveOpen(false);
-                    closeMenu();
-                    onMoveToFolder(chat, folder.id);
-                  }}
-                  className={cx(
-                    "chat-history-menu-item hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none",
-                    chat.folder_id === folder.id ? "text-white" : "text-neutral-300",
-                  )}
-                >
-                  <MaskIcon src="/icons/folder.png" size={13} />
-                  <span className="truncate">{folder.name}</span>
-                </button>
-              ))}
-              {chat.folder_id && (
-                <button
-                  type="button"
-                  role="menuitem"
-                  onClick={() => {
-                    setMoveOpen(false);
-                    closeMenu();
-                    onMoveToFolder(chat, null);
-                  }}
-                  className="chat-history-menu-item text-neutral-300 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-                >
-                  <X size={13} />
-                  Remove from folder
-                </button>
-              )}
-            </div>
-          )}
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onExport(chat);
-            }}
-            className="chat-history-menu-item text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-          >
-            <i className="fi fi-rr-file-export text-[14px] leading-none" aria-hidden="true" />
-            Export chat
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onDelete(chat.id);
-            }}
-            className="chat-history-menu-item text-red-300 hover:bg-red-500/10 focus:bg-red-500/10 focus:outline-none"
-          >
-            <Trash2 size={14} />
-            Delete chat
-          </button>
-        </>
-      )}
-    </OverflowActions>
-  );
-}
-
-function FolderActions({ folder, onRename, onDelete, onNewChat }) {
-  return (
-    <OverflowActions
-      id={`folder-${folder.id}`}
-      title={folder.name}
-      label="Folder actions"
-      menuWidth={196}
-    >
-      {(closeMenu) => (
-        <>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onNewChat(folder.id);
-            }}
-            className="chat-history-menu-item text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-          >
-            <MaskIcon src="/icons/new-message.png" size={14} />
-            New chat here
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onRename(folder);
-            }}
-            className="chat-history-menu-item text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-          >
-            <Pencil size={14} />
-            Edit name
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onDelete(folder);
-            }}
-            className="chat-history-menu-item text-red-300 hover:bg-red-500/10 focus:bg-red-500/10 focus:outline-none"
-          >
-            <Trash2 size={14} />
-            Delete folder
-          </button>
-        </>
-      )}
-    </OverflowActions>
-  );
-}
-
-function StoryHistoryActions({ story, onRename, onExport, onDelete, exportDisabled = false }) {
-  return (
-    <OverflowActions
-      id={`story-${story.id}`}
-      title={story.title}
-      label="Story actions"
-      menuWidth={164}
-    >
-      {(closeMenu) => (
-        <>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onRename(story);
-            }}
-            className="chat-history-menu-item text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-          >
-            <Pencil size={14} />
-            Edit name
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            disabled={exportDisabled}
-            title={exportDisabled ? "Wait for writing to finish" : undefined}
-            onClick={() => {
-              if (exportDisabled) return;
-              closeMenu();
-              onExport(story);
-            }}
-            className={cx(
-              "chat-history-menu-item focus:outline-none",
-              exportDisabled
-                ? "cursor-not-allowed text-neutral-600"
-                : "text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07]",
-            )}
-          >
-            <i className="fi fi-rr-file-export text-[14px] leading-none" aria-hidden="true" />
-            Export
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onDelete(story);
-            }}
-            className="chat-history-menu-item text-red-300 hover:bg-red-500/10 focus:bg-red-500/10 focus:outline-none"
-          >
-            <Trash2 size={14} />
-            Delete story
-          </button>
-        </>
-      )}
-    </OverflowActions>
-  );
-}
-
-function ChapterHistoryActions({ chapter, onRename, onDelete, onToggleContext }) {
-  return (
-    <OverflowActions
-      id={`chapter-${chapter.id}`}
-      title={chapter.title}
-      label="Chapter actions"
-      menuWidth={164}
-    >
-      {(closeMenu) => (
-        <>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onRename(chapter);
-            }}
-            className="chat-history-menu-item text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-          >
-            <Pencil size={14} />
-            Edit name
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onToggleContext(chapter);
-            }}
-            className="chat-history-menu-item text-neutral-200 hover:bg-white/[0.07] focus:bg-white/[0.07] focus:outline-none"
-          >
-            {chapter.disabled ? <Eye size={14} /> : <EyeOff size={14} />}
-            {chapter.disabled ? "Show chapter" : "Hide chapter"}
-          </button>
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              closeMenu();
-              onDelete(chapter);
-            }}
-            className="chat-history-menu-item text-red-300 hover:bg-red-500/10 focus:bg-red-500/10 focus:outline-none"
-          >
-            <Trash2 size={14} />
-            Delete chapter
-          </button>
-        </>
-      )}
-    </OverflowActions>
-  );
-}
-
-function SidebarSearchModal({
-  open,
-  items,
-  onSelect,
-  onClose,
-  label = "Search chats",
-  emptyText = "Your conversations will appear here.",
-  noMatchText = "No chats match that search.",
-  fallbackTitle = "Untitled chat",
-}) {
-  const [rendered, setRendered] = useState(open);
-  const [phase, setPhase] = useState(open ? "open" : "closed");
-  const [query, setQuery] = useState("");
-  const inputRef = useRef(null);
-
-  useEffect(() => {
-    if (open) {
-      setRendered(true);
-      setPhase("open");
-      setQuery("");
-      requestAnimationFrame(() => inputRef.current?.focus());
-      return undefined;
-    }
-
-    if (!rendered) return undefined;
-
-    setPhase("closing");
-    const closeMs =
-      parseFloat(
-        getComputedStyle(document.documentElement).getPropertyValue("--modal-close-dur"),
-      ) || 150;
-    const timeoutId = window.setTimeout(() => {
-      setRendered(false);
-      setPhase("closed");
-    }, closeMs);
-    return () => window.clearTimeout(timeoutId);
-  }, [open, rendered]);
-
-  useEffect(() => {
-    if (!rendered) return undefined;
-
-    function handleKeyDown(event) {
-      if (event.key === "Escape") onClose();
-    }
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [onClose, rendered]);
-
-  const recentItems = useMemo(() => {
-    const sorted = [...items].sort((first, second) => {
-      const firstTime = Date.parse(first.updated_at || first.created_at || "") || 0;
-      const secondTime = Date.parse(second.updated_at || second.created_at || "") || 0;
-      return secondTime - firstTime;
-    });
-
-    const trimmed = query.trim().toLowerCase();
-    if (!trimmed) return sorted;
-
-    return sorted.filter((item) => (item.title || "").toLowerCase().includes(trimmed));
-  }, [items, query]);
-
-  if (!rendered) return null;
-
-  const isOpen = phase === "open";
-
-  return createPortal(
-    <div className="fixed inset-0 z-[80] grid place-items-center px-4 py-6">
-      <button
-        type="button"
-        aria-label="Close search"
-        className={cx(
-          "absolute inset-0 bg-black/60 backdrop-blur-sm transition-[opacity,backdrop-filter] duration-150 ease-out",
-          isOpen ? "opacity-100" : "opacity-0",
-        )}
-        onClick={onClose}
-      />
-      <section
-        role="dialog"
-        aria-modal="true"
-        aria-label={label}
-        className={cx(
-          "t-modal relative z-10 flex max-h-[min(620px,calc(100dvh-3rem))] w-full max-w-[560px] flex-col overflow-hidden rounded-[26px] bg-[#191919] text-neutral-100 [box-shadow:var(--shadow-surface)]",
-          isOpen ? "is-open" : "is-closing",
-        )}
-      >
-        <header className="flex shrink-0 items-center gap-4 px-6 pb-2 pt-5">
-          <input
-            ref={inputRef}
-            type="text"
-            value={query}
-            onChange={(event) => setQuery(event.target.value)}
-            placeholder="Search..."
-            data-1p-ignore="true"
-            className="min-w-0 flex-1 bg-transparent text-[19px] font-normal leading-7 text-neutral-100 placeholder:text-neutral-500 focus:outline-none"
-          />
-          <button
-            type="button"
-            aria-label="Close search"
-            title="Close search"
-            onClick={onClose}
-            className={cx(
-              "inline-flex h-9 w-9 shrink-0 items-center justify-center rounded-lg text-neutral-400 hover:text-white focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50",
-              CONTROL_MOTION,
-            )}
-          >
-            <X size={20} />
-          </button>
-        </header>
-
-        <div className="chat-rail-scrollbar min-h-0 flex-1 overflow-y-auto px-3 pb-4 pt-2">
-          {recentItems.length > 0 && (
-            <div className="px-3 pb-1 pt-2 text-[15px] leading-6 text-neutral-500">
-              {query.trim() ? "Results" : "Last opened"}
-            </div>
-          )}
-
-          {recentItems.length === 0 ? (
-            <div className="px-3 py-6 text-sm leading-6 text-neutral-500">
-              {query.trim() ? noMatchText : emptyText}
-            </div>
-          ) : (
-            <div className="space-y-0.5">
-              {recentItems.map((item) => (
-                <button
-                  key={item.id}
-                  type="button"
-                  onClick={() => onSelect(item.id)}
-                  className={cx(
-                    "flex w-full items-center rounded-xl px-3 py-2.5 text-left text-[15px] leading-6 text-neutral-100 hover:bg-white/[0.06] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45",
-                    CONTROL_MOTION,
-                  )}
-                >
-                  <span className="truncate">{item.title || fallbackTitle}</span>
-                </button>
-              ))}
-            </div>
-          )}
-        </div>
-      </section>
-    </div>,
-    document.body,
-  );
-}
-
-function ChangelogModal({ open, onClose }) {
-  const [rendered, setRendered] = useState(open);
-  const [phase, setPhase] = useState(open ? "open" : "closed");
-  const [status, setStatus] = useState("idle");
-  const [release, setRelease] = useState(null);
-  const panelRef = useRef(null);
-
-  useEffect(() => {
-    if (open) {
-      setRendered(true);
-      setPhase("open");
-      requestAnimationFrame(() => panelRef.current?.focus());
-      return undefined;
-    }
-
-    if (!rendered) return undefined;
-
-    setPhase("closing");
-    const closeMs =
-      parseFloat(
-        getComputedStyle(document.documentElement).getPropertyValue("--modal-close-dur"),
-      ) || 150;
-    const timeoutId = window.setTimeout(() => {
-      setRendered(false);
-      setPhase("closed");
-    }, closeMs);
-    return () => window.clearTimeout(timeoutId);
-  }, [open, rendered]);
-
-  useEffect(() => {
-    if (!rendered) return undefined;
-
-    function handleKeyDown(event) {
-      if (event.key === "Escape") onClose();
-    }
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [onClose, rendered]);
-
-  useEffect(() => {
-    if (!open) return undefined;
-
-    let cancelled = false;
-    setStatus("loading");
-
-    fetch(GITHUB_RELEASES_LATEST_URL, {
-      headers: { Accept: "application/vnd.github+json" },
-    })
-      .then((response) => {
-        if (!response.ok) throw new Error(`GitHub returned ${response.status}`);
-        return response.json();
-      })
-      .then((data) => {
-        if (cancelled) return;
-        setRelease(data);
-        setStatus("ready");
-      })
-      .catch(() => {
-        if (cancelled) return;
-        setStatus("error");
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [open]);
-
-  if (!rendered) return null;
-
-  const isOpen = phase === "open";
-  const releaseBody = release?.body;
-
-  return createPortal(
-    <div className="fixed inset-0 z-[80] grid place-items-center px-4 py-6">
-      <button
-        type="button"
-        aria-label="Close changelog"
-        className={cx(
-          "absolute inset-0 bg-black/60 backdrop-blur-sm transition-[opacity,backdrop-filter] duration-150 ease-out",
-          isOpen ? "opacity-100" : "opacity-0",
-        )}
-        onClick={onClose}
-      />
-      <section
-        ref={panelRef}
-        role="dialog"
-        aria-modal="true"
-        aria-label="Changelog"
-        tabIndex={-1}
-        className={cx(
-          "t-modal relative z-10 flex max-h-[min(640px,calc(100dvh-2rem))] w-full max-w-[560px] flex-col overflow-hidden rounded-[26px] bg-[#191919] text-neutral-100 outline-none [box-shadow:var(--shadow-surface)]",
-          isOpen ? "is-open" : "is-closing",
-        )}
-      >
-        <button
-          type="button"
-          onClick={onClose}
-          className={cx(
-            "absolute right-3 top-3 z-10 grid h-10 w-10 shrink-0 place-items-center rounded-full text-neutral-400 hover:text-white focus:outline-none focus-visible:ring-2 focus-visible:ring-white/20 sm:right-4 sm:top-4",
-            CONTROL_MOTION,
-          )}
-          aria-label="Close changelog"
-        >
-          <X size={17} />
-        </button>
-
-        <div className="min-h-0 flex-1 overflow-y-auto px-5 pb-6 pt-5 sm:px-6 sm:pt-6">
-          {status === "loading" && (
-            <p className="text-sm leading-5 text-neutral-500">Loading latest release…</p>
-          )}
-          {status === "error" && (
-            <p className="text-sm leading-5 text-neutral-500">
-              Couldn't load the changelog right now. Try again later.
-            </p>
-          )}
-          {status === "ready" && (
-            <div className="text-[13px] leading-6 text-neutral-300">
-              <MarkdownContent>
-                {releaseBody || "No description provided for this release."}
-              </MarkdownContent>
-            </div>
-          )}
-        </div>
-      </section>
-    </div>,
-    document.body,
-  );
-}
-
-let changelogAutoCheckStarted = false;
-
-function FeedbackLink() {
-  const [changelogOpen, setChangelogOpen] = useState(false);
-
-  useEffect(() => {
-    if (changelogAutoCheckStarted) return;
-    changelogAutoCheckStarted = true;
-
-    api("/api/changelog/status")
-      .then((status) => {
-        if (!status?.should_show) return;
-        setChangelogOpen(true);
-        return api("/api/changelog/seen", { method: "POST" });
-      })
-      .catch(() => {});
-  }, []);
-
-  return (
-    <>
-      <div className="flex w-full items-center justify-center gap-2">
-        <button
-          type="button"
-          onClick={() => setChangelogOpen(true)}
-          className={cx(
-            "flex h-9 flex-1 items-center justify-center rounded-xl px-2 text-[13px] font-medium text-neutral-500 hover:bg-white/[0.045] hover:text-neutral-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45",
-            CONTROL_MOTION,
-          )}
-        >
-          Changelog
-        </button>
-        <a
-          href={FEEDBACK_FORM_URL}
-          target="_blank"
-          rel="noreferrer"
-          className={cx(
-            "flex h-9 flex-1 items-center justify-center rounded-xl px-2 text-[13px] font-medium text-neutral-500 hover:bg-white/[0.045] hover:text-neutral-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45",
-            CONTROL_MOTION,
-          )}
-        >
-          Feedback
-        </a>
-      </div>
-      <ChangelogModal open={changelogOpen} onClose={() => setChangelogOpen(false)} />
-    </>
-  );
-}
-
-function SidebarGroup({ label, open, onToggle, children, dropProps = {}, dropActive = false }) {
-  const panelId = `${label.toLowerCase()}-chat-history`;
-
-  return (
-    <section
-      className={cx(
-        "t-acc chat-history-group rounded-2xl border border-transparent px-1 transition-[background-color,border-color,box-shadow] duration-150 ease-out",
-        dropActive && "border-white/20 bg-white/[0.07]",
-      )}
-      data-open={String(open)}
-      {...dropProps}
-    >
-      <button
-        type="button"
-        className="t-acc-head chat-history-group-heading"
-        aria-expanded={open}
-        aria-controls={panelId}
-        onClick={onToggle}
-      >
-        <span>{label}</span>
-        <span className="t-acc-chevron chat-history-group-chevron" aria-hidden="true">
-          <svg viewBox="0 0 16 16" fill="none">
-            <path d="M4 6.5L8 10.5L12 6.5" stroke="currentColor" strokeWidth="1.75" strokeLinecap="round" strokeLinejoin="round" />
-          </svg>
-        </span>
-      </button>
-      <div id={panelId} className="t-acc-panel">
-        <div className="t-acc-panel-inner chat-history-group-items">
-          {children}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function ConversationRail({
-  chats,
-  activeChatId,
-  models,
-  onNewChat,
-  onLoadChat,
-  onRenameChat,
-  onDeleteChat,
-  onExportChat,
-  onTogglePinChat,
-  namingChatId,
-  folders,
-  onCreateFolder,
-  onRenameFolder,
-  onDeleteFolder,
-  onMoveChatToFolder,
-  onNewChatInFolder,
-  mobileOpen,
-  onCloseMobile,
-  collapsed,
-  onCollapse,
-  highlightFirstChatActions,
-  chatMode,
-  previousChatMode,
-  onChatModeChange,
-}) {
-  const [railScrolling, setRailScrolling] = useState(false);
-  const [railScrolled, setRailScrolled] = useState(false);
-  const [railHasMoreBelow, setRailHasMoreBelow] = useState(false);
-  const [renamingChatId, setRenamingChatId] = useState(null);
-  const [renameDraft, setRenameDraft] = useState("");
-  const [pinnedOpen, setPinnedOpen] = useState(true);
-  const [recentsOpen, setRecentsOpen] = useState(true);
-  const [foldersOpen, setFoldersOpen] = useState(true);
-  const [searchOpen, setSearchOpen] = useState(false);
-  const [newFolderOpen, setNewFolderOpen] = useState(false);
-  const [renamingFolderId, setRenamingFolderId] = useState(null);
-  const [folderRenameDraft, setFolderRenameDraft] = useState("");
-  const [expandedFolderIds, setExpandedFolderIds] = useState([]);
-  const [dragChatId, setDragChatId] = useState(null);
-  const [dropFolderId, setDropFolderId] = useState(null);
-  const [recentsDropActive, setRecentsDropActive] = useState(false);
-  const railScrollTimeoutRef = useRef(null);
-  const railRef = useRef(null);
-  const skipRenameCommitRef = useRef(false);
-  const skipFolderRenameCommitRef = useRef(false);
-
-  useEffect(
-    () => () => {
-      if (railScrollTimeoutRef.current) {
-        window.clearTimeout(railScrollTimeoutRef.current);
-      }
-    },
-    [],
-  );
-
-  function updateRailEdges(element) {
-    const bottomOffset = element.scrollHeight - element.clientHeight - element.scrollTop;
-    setRailScrolled(element.scrollTop > 2);
-    setRailHasMoreBelow(bottomOffset > 2);
-  }
-
-  useEffect(() => {
-    const rail = railRef.current;
-    if (!rail) return;
-
-    const frameId = requestAnimationFrame(() => updateRailEdges(rail));
-    const resizeObserver = typeof ResizeObserver === "undefined"
-      ? null
-      : new ResizeObserver(() => updateRailEdges(rail));
-    resizeObserver?.observe(rail);
-
-    return () => {
-      cancelAnimationFrame(frameId);
-      resizeObserver?.disconnect();
-    };
-  }, [chats.length, folders.length, pinnedOpen, recentsOpen, foldersOpen, expandedFolderIds]);
-
-  function handleRailScroll(event) {
-    updateRailEdges(event.currentTarget);
-    setRailScrolling(true);
-    window.clearTimeout(railScrollTimeoutRef.current);
-    railScrollTimeoutRef.current = window.setTimeout(
-      () => setRailScrolling(false),
-      650,
-    );
-  }
-
-  function startRename(chat) {
-    if (chat.id === namingChatId) return;
-    skipRenameCommitRef.current = false;
-    setRenamingChatId(chat.id);
-    setRenameDraft(chat.title || "");
-  }
-
-  function cancelRename() {
-    skipRenameCommitRef.current = true;
-    setRenamingChatId(null);
-    setRenameDraft("");
-  }
-
-  async function commitRename(chat) {
-    if (skipRenameCommitRef.current) {
-      skipRenameCommitRef.current = false;
-      return;
-    }
-    const nextTitle = renameDraft.trim();
-    if (!nextTitle || nextTitle === chat.title) {
-      setRenamingChatId(null);
-      setRenameDraft("");
-      return;
-    }
-    await onRenameChat(chat.id, nextTitle);
-    setRenamingChatId(null);
-    setRenameDraft("");
-  }
-
-  async function createFolder(name) {
-    const folder = await onCreateFolder(name);
-    setFoldersOpen(true);
-    if (folder?.id) setExpandedFolderIds((current) => [...current, folder.id]);
-  }
-
-  function toggleFolderExpanded(folderId) {
-    setExpandedFolderIds((current) => (
-      current.includes(folderId)
-        ? current.filter((id) => id !== folderId)
-        : [...current, folderId]
-    ));
-  }
-
-  function startFolderRename(folder) {
-    skipFolderRenameCommitRef.current = false;
-    setRenamingFolderId(folder.id);
-    setFolderRenameDraft(folder.name || "");
-  }
-
-  function cancelFolderRename() {
-    skipFolderRenameCommitRef.current = true;
-    setRenamingFolderId(null);
-    setFolderRenameDraft("");
-  }
-
-  async function commitFolderRename(folder) {
-    if (skipFolderRenameCommitRef.current) {
-      skipFolderRenameCommitRef.current = false;
-      return;
-    }
-    const nextName = folderRenameDraft.trim();
-    setRenamingFolderId(null);
-    setFolderRenameDraft("");
-    if (nextName && nextName !== folder.name) {
-      await onRenameFolder(folder.id, nextName);
-    }
-  }
-
-  function handleChatDrop(folderId) {
-    const chat = chats.find((item) => item.id === dragChatId);
-    setDragChatId(null);
-    setDropFolderId(null);
-    setRecentsDropActive(false);
-    if (!chat) return;
-    if (folderId) setExpandedFolderIds((current) => (
-      current.includes(folderId) ? current : [...current, folderId]
-    ));
-    void onMoveChatToFolder(chat, folderId);
-  }
-
-  const pinnedChats = chats.filter((chat) => chat.pinned);
-  const recentChats = chats.filter((chat) => !chat.pinned && !chat.folder_id);
-  const draggedFolderChat = chats.find((chat) => chat.id === dragChatId && chat.folder_id);
-
-  function renderChatRows(groupChats, startIndex = 0) {
-    return groupChats.map((chat, chatIndex) => {
-      const renaming = renamingChatId === chat.id;
-      const index = startIndex + chatIndex;
-
-      return (
-        <div
-          key={chat.id}
-          draggable={!renaming}
-          onDragStart={(event) => {
-            event.dataTransfer.effectAllowed = "move";
-            event.dataTransfer.setData("text/plain", chat.id);
-            setDragChatId(chat.id);
-          }}
-          onDragEnd={() => {
-            setDragChatId(null);
-            setDropFolderId(null);
-            setRecentsDropActive(false);
-          }}
-          className={cx(
-            "group relative grid select-none grid-cols-[minmax(0,1fr)_auto] items-center gap-1 rounded-2xl border border-transparent px-2 py-1 transition-[background-color,border-color,box-shadow] duration-150 ease-out",
-            chat.id === activeChatId
-              ? "bg-white/[0.08] shadow-[var(--shadow-border)]"
-              : "hover:bg-white/[0.045] hover:shadow-[var(--shadow-border)]",
-            dragChatId === chat.id && "opacity-45",
-          )}
-        >
-          {renaming ? (
-            <div className="min-h-8 min-w-0 rounded-xl px-1 py-0.5">
-              <input
-                autoFocus
-                value={renameDraft}
-                onChange={(event) => setRenameDraft(event.target.value)}
-                data-1p-ignore="true"
-                onBlur={() => commitRename(chat)}
-                onKeyDown={(event) => {
-                  if (event.key === "Enter") event.currentTarget.blur();
-                  if (event.key === "Escape") {
-                    event.preventDefault();
-                    cancelRename();
-                  }
-                }}
-                className="block h-6 w-full min-w-0 rounded-md bg-white/[0.06] px-1.5 text-sm font-medium text-neutral-100 outline-none shadow-[var(--shadow-border)]"
-              />
-              <div className="truncate text-[11px] leading-3 text-neutral-500">
-                {promptModelName(models, chat.model)}
-              </div>
-            </div>
-          ) : (
-            <button
-              type="button"
-              onClick={() => {
-                onLoadChat(chat.id);
-                onCloseMobile();
-              }}
-              className={cx(
-                "min-h-8 min-w-0 rounded-xl px-1 py-0.5 text-left focus:outline-none focus-visible:ring-2 focus-visible:ring-white/15",
-                CONTROL_MOTION,
-              )}
-            >
-              <div className="truncate text-balance text-sm font-medium leading-4 text-neutral-100">
-                {chat.title}
-              </div>
-              <div className="truncate text-[11px] leading-3 text-neutral-500">
-                {promptModelName(models, chat.model)}
-              </div>
-            </button>
-          )}
-          <ChatHistoryActions
-            chat={chat}
-            isFirst={index === 0}
-            forceVisible={index === 0 && highlightFirstChatActions}
-            folders={folders}
-            renameDisabled={chat.id === namingChatId}
-            onRename={startRename}
-            onDelete={onDeleteChat}
-            onExport={onExportChat}
-            onTogglePin={onTogglePinChat}
-            onMoveToFolder={onMoveChatToFolder}
-          />
-        </div>
-      );
-    });
-  }
-
-  function renderFolderRows() {
-    return folders.map((folder) => {
-      const renaming = renamingFolderId === folder.id;
-      const expanded = expandedFolderIds.includes(folder.id);
-      const folderChats = chats.filter((chat) => chat.folder_id === folder.id);
-
-      return (
-        <div key={folder.id}>
-        <div
-          onDragOver={(event) => {
-            if (!dragChatId) return;
-            event.preventDefault();
-            event.dataTransfer.dropEffect = "move";
-            setDropFolderId(folder.id);
-          }}
-          onDragLeave={() => setDropFolderId((current) => (current === folder.id ? null : current))}
-          onDrop={(event) => {
-            event.preventDefault();
-            handleChatDrop(folder.id);
-          }}
-          className={cx(
-            "group relative grid select-none grid-cols-[minmax(0,1fr)_auto] items-center gap-1 rounded-2xl border border-transparent px-2 py-1 transition-[background-color,border-color,box-shadow] duration-150 ease-out hover:bg-white/[0.045] hover:shadow-[var(--shadow-border)]",
-            dropFolderId === folder.id && "border-white/20 bg-white/[0.07]",
-          )}
-        >
-          {renaming ? (
-            <div className="flex min-h-8 min-w-0 items-center gap-2 rounded-xl px-1 py-0.5">
-              <MaskIcon src="/icons/folder.png" size={16} className="text-neutral-300" />
-              <input
-                autoFocus
-                value={folderRenameDraft}
-                onChange={(event) => setFolderRenameDraft(event.target.value)}
-                data-1p-ignore="true"
-                onBlur={() => commitFolderRename(folder)}
-                onKeyDown={(event) => {
-                  if (event.key === "Enter") event.currentTarget.blur();
-                  if (event.key === "Escape") {
-                    event.preventDefault();
-                    cancelFolderRename();
-                  }
-                }}
-                className="block h-6 w-full min-w-0 rounded-md bg-white/[0.06] px-1.5 text-sm font-medium text-neutral-100 outline-none shadow-[var(--shadow-border)]"
-              />
-            </div>
-          ) : (
-            <button
-              type="button"
-              aria-expanded={expanded}
-              onClick={() => toggleFolderExpanded(folder.id)}
-              className={cx(
-                "flex min-h-8 min-w-0 items-center gap-2 rounded-xl px-1 py-0.5 text-left focus:outline-none focus-visible:ring-2 focus-visible:ring-white/15",
-                CONTROL_MOTION,
-              )}
-            >
-              <MaskIcon src="/icons/folder.png" size={16} className="text-neutral-300" />
-              <span className="truncate text-sm font-medium leading-4 text-neutral-100">
-                {folder.name}
-              </span>
-              {folderChats.length > 0 && (
-                <span className="shrink-0 text-[11px] leading-3 text-neutral-500">
-                  {folderChats.length}
-                </span>
-              )}
-            </button>
-          )}
-          <FolderActions
-            folder={folder}
-            onRename={startFolderRename}
-            onDelete={onDeleteFolder}
-            onNewChat={onNewChatInFolder}
-          />
-        </div>
-
-        {expanded && (
-          <div className="mt-1 space-y-1 border-l border-white/10 pl-2">
-            {folderChats.length > 0 ? (
-              renderChatRows(folderChats, 1)
-            ) : (
-              <div className="px-2 py-2 text-[12px] leading-4 text-neutral-500">
-                Drag a chat here, or use New chat in this folder.
-              </div>
-            )}
-          </div>
-        )}
-        </div>
-      );
-    });
-  }
-
-  function renderHistoryGroup(label, groupChats, open, onToggle, startIndex = 0) {
-    return renderGroupShell(label, open, onToggle, renderChatRows(groupChats, startIndex));
-  }
-
-  function renderGroupShell(label, open, onToggle, children, dropProps = {}, dropActive = false) {
-    return (
-      <SidebarGroup
-        label={label}
-        open={open}
-        onToggle={onToggle}
-        dropProps={dropProps}
-        dropActive={dropActive}
-      >
-        {children}
-      </SidebarGroup>
-    );
-  }
-
-  const folderGroup = folders.length > 0 && renderGroupShell(
-    "Folders",
-    foldersOpen,
-    () => setFoldersOpen((current) => !current),
-    renderFolderRows(),
-  );
-
-  const historyItems =
-    chats.length === 0 ? (
-      <div className="space-y-4">
-        {folderGroup}
-        <div className="px-3 py-8 text-pretty text-sm leading-6 text-neutral-500">
-          {chatMode === "write"
-            ? "Your stories will appear here."
-            : "Your conversations will appear here."}
-        </div>
-      </div>
-    ) : (
-      <div className="space-y-4">
-        {pinnedChats.length > 0 && renderHistoryGroup(
-          "Pinned",
-          pinnedChats,
-          pinnedOpen,
-          () => setPinnedOpen((current) => !current),
-        )}
-        {folderGroup}
-        {renderGroupShell(
-          "Recents",
-          recentsOpen,
-          () => setRecentsOpen((current) => !current),
-          renderChatRows(recentChats, pinnedChats.length),
-          {
-            onDragOver: (event) => {
-              if (!draggedFolderChat) return;
-              event.preventDefault();
-              event.dataTransfer.dropEffect = "move";
-              setRecentsDropActive(true);
-            },
-            onDragLeave: () => setRecentsDropActive(false),
-            onDrop: (event) => {
-              event.preventDefault();
-              setRecentsDropActive(false);
-              handleChatDrop(null);
-            },
-          },
-          recentsDropActive,
-        )}
-      </div>
-    );
-
-  return (
-    <>
-      <div
-        className={cx(
-          "fixed inset-0 z-30 bg-black/55 opacity-0 backdrop-blur-sm transition-[opacity,backdrop-filter] duration-200 ease-out lg:hidden",
-          mobileOpen ? "pointer-events-auto opacity-100" : "pointer-events-none",
-        )}
-        onClick={onCloseMobile}
-      />
-      <aside
-        className={cx(
-          "chat-sidebar t-resize fixed inset-y-0 left-0 z-40 flex w-[292px] flex-col overflow-hidden border-r border-line bg-[#080808] lg:static lg:z-auto lg:translate-x-0",
-          collapsed
-            ? "lg:w-0 lg:-translate-x-3 lg:border-r-0 lg:border-transparent lg:opacity-0"
-            : "lg:w-[276px] lg:opacity-100",
-          mobileOpen ? "translate-x-0" : "-translate-x-full",
-        )}
-      >
-        <div
-          className={cx(
-            "chat-sidebar-content flex h-full w-[292px] flex-col p-4 lg:w-[276px]",
-            collapsed
-              ? "lg:-translate-x-8 lg:opacity-0"
-              : "lg:translate-x-0 lg:opacity-100",
-          )}
-        >
-          <div className="mb-4 flex items-center justify-between gap-2 pl-2">
-            <div className="flex min-w-0 items-baseline gap-1.5">
-              <span className="truncate text-[19px] font-bold tracking-[-0.015em] text-white">
-                RouterChat
-              </span>
-              <span className="shrink-0 text-[19px] font-bold tracking-[-0.015em] text-neutral-500">
-                {APP_VERSION}
-              </span>
-            </div>
-
-            <div className="flex shrink-0 items-center">
-              <button
-                type="button"
-                aria-label="Search chats"
-                title="Search chats"
-                onClick={() => setSearchOpen(true)}
-                className={cx(
-                  "hidden h-10 w-10 shrink-0 items-center justify-center rounded-lg text-neutral-400 hover:bg-white/[0.08] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50 lg:inline-flex",
-                  CONTROL_MOTION,
-                )}
-              >
-                <MaskIcon src="/icons/search.png" size={19} />
-              </button>
-              <button
-                type="button"
-                aria-label="Collapse sidebar"
-                title="Collapse sidebar"
-                data-tour="collapse-sidebar-button"
-                onClick={onCollapse}
-                className={cx(
-                  "hidden h-10 w-10 shrink-0 items-center justify-center rounded-lg text-neutral-400 hover:bg-white/[0.08] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50 lg:inline-flex",
-                  CONTROL_MOTION,
-                )}
-              >
-                <MaskIcon src="/icons/sidebar.png" size={15.5} />
-              </button>
-              <button
-                type="button"
-                aria-label="Close chats"
-                title="Close chats"
-                onClick={onCloseMobile}
-                className={cx(
-                  "inline-flex h-10 w-9 shrink-0 items-center justify-center rounded-lg text-neutral-400 hover:text-white focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50 lg:hidden",
-                  CONTROL_MOTION,
-                )}
-              >
-                <X size={19} />
-              </button>
-            </div>
-          </div>
-
-          <div className="mb-3.5 flex justify-center">
-            <SlidingTabs
-              options={CHAT_MODES}
-              value={chatMode}
-              fromValue={previousChatMode}
-              onChange={onChatModeChange}
-              getValue={(mode) => mode.value}
-              getLabel={(mode) => mode.label}
-              ariaLabel="Interaction mode"
-              className="sidebar-mode-tabs"
-            />
-          </div>
-
-          <div>
-            <button
-              type="button"
-              onClick={() => {
-                onNewChat();
-                onCloseMobile();
-              }}
-              className={cx(
-                "flex h-9 w-full items-center gap-3 rounded-xl bg-transparent px-2 text-[15px] font-medium text-white hover:bg-white/[0.06] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45",
-                CONTROL_MOTION,
-              )}
-            >
-              <MaskIcon src="/icons/new-message.png" size={20} className="text-neutral-200" />
-              New chat
-            </button>
-
-            <button
-              type="button"
-              onClick={() => setNewFolderOpen(true)}
-              className={cx(
-                "flex h-9 w-full items-center gap-3 rounded-xl bg-transparent px-2 text-[15px] font-medium text-white hover:bg-white/[0.06] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45",
-                CONTROL_MOTION,
-              )}
-            >
-              <MaskIcon src="/icons/folder.png" size={20} className="text-neutral-200" />
-              New folder
-            </button>
-          </div>
-
-          <div className="relative min-h-0 flex-1">
-            <nav
-              ref={railRef}
-              onScroll={handleRailScroll}
-              className={cx(
-                "chat-rail-scrollbar h-full space-y-1 overflow-y-auto pr-1",
-                railScrolling && "is-scrolling",
-              )}
-            >
-              {historyItems}
-            </nav>
-            <div
-              aria-hidden="true"
-              className={cx(
-                "sidebar-list-fade pointer-events-none absolute inset-x-0 top-0 h-6 bg-gradient-to-b from-[#080808]/95 to-transparent transition-opacity duration-150 ease-out",
-                railScrolled ? "opacity-100" : "opacity-0",
-              )}
-            />
-            <div
-              aria-hidden="true"
-              className={cx(
-                "sidebar-list-fade pointer-events-none absolute inset-x-0 bottom-0 h-7 bg-gradient-to-t from-[#080808]/95 to-transparent transition-opacity duration-150 ease-out",
-                railHasMoreBelow ? "opacity-100" : "opacity-0",
-              )}
-            />
-          </div>
-
-          <footer className="mt-1 -mb-2">
-            <FeedbackLink />
-          </footer>
-        </div>
-      </aside>
-
-      <SidebarSearchModal
-        open={searchOpen}
-        items={chats}
-        onSelect={(chatId) => {
-          setSearchOpen(false);
-          onLoadChat(chatId);
-          onCloseMobile();
-        }}
-        onClose={() => setSearchOpen(false)}
-      />
-
-      <NamePromptModal
-        open={newFolderOpen}
-        onClose={() => setNewFolderOpen(false)}
-        onCreate={createFolder}
-        heading="Name your new folder"
-        description="You can rename your folder from the sidebar at any time"
-        placeholder="Folder name"
-        inputLabel="Folder name"
-        submitLabel="Create folder"
-        dialogLabel="Close new folder dialog"
-      />
-    </>
-  );
-}
-
-function SidebarRevealButton({ visible, onClick }) {
-  const [nearEdge, setNearEdge] = useState(false);
-
-  if (!visible) return null;
-
-  return (
-    <div
-      className="fixed inset-y-0 left-0 z-50 hidden w-16 lg:block"
-      onMouseEnter={() => setNearEdge(true)}
-      onMouseLeave={() => setNearEdge(false)}
-    >
-      <button
-        type="button"
-        aria-label="Expand sidebar"
-        title="Expand sidebar"
-        onClick={onClick}
-        onFocus={() => setNearEdge(true)}
-        onBlur={() => setNearEdge(false)}
-        className={cx(
-          "sidebar-reveal-button absolute left-3 top-1/2 inline-flex h-11 w-11 -translate-y-1/2 items-center justify-center rounded-full border border-line bg-[#181818]/95 text-neutral-300 [box-shadow:var(--shadow-surface)] backdrop-blur-xl hover:border-white/15 hover:bg-white/[0.09] hover:text-white focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50",
-          nearEdge
-            ? "translate-x-0 opacity-100"
-            : "-translate-x-2 opacity-0 focus-visible:translate-x-0 focus-visible:opacity-100",
-        )}
-      >
-        <PanelLeftOpen size={18} />
-      </button>
-    </div>
-  );
-}
-
-function PromptNavigationRail({ messages, streamRef, visible, activeChatId }) {
-  const railRef = useRef(null);
-  const frameRef = useRef(null);
-  const [railItems, setRailItems] = useState([]);
-  const [activePromptId, setActivePromptId] = useState(null);
-  const [previewPromptId, setPreviewPromptId] = useState(null);
-
-  const promptMessages = useMemo(
-    () => messages.filter((message) => message.role === "user"),
-    [messages],
-  );
-
-  const measureRail = useCallback(() => {
-    frameRef.current = null;
-
-    const scroller = streamRef.current;
-    const rail = railRef.current;
-    if (!visible || !activeChatId || !scroller || !rail || promptMessages.length === 0) {
-      setRailItems([]);
-      setActivePromptId(null);
-      return;
-    }
-
-    const railPadding = 12;
-    const railHeight = Math.max(rail.clientHeight - railPadding * 2, 1);
-    const railCenter = railPadding + railHeight / 2;
-    const promptGap = 8;
-    const maxScroll = Math.max(scroller.scrollHeight - scroller.clientHeight, 1);
-    const activeLine = scroller.scrollTop + scroller.clientHeight * 0.5;
-    const scrollerTop = scroller.getBoundingClientRect().top;
-
-    const messageNodes = Array.from(scroller.querySelectorAll("[data-message-id]"));
-
-    const baseItems = promptMessages.map((message, index) => {
-      const node = messageNodes.find(
-        (messageNode) => messageNode.dataset.messageId === String(message.id),
-      );
-      if (!node) return null;
-
-      const nodeTop = node.getBoundingClientRect().top;
-      const scrollTop = Math.min(
-        Math.max(scroller.scrollTop + nodeTop - scrollerTop - 24, 0),
-        maxScroll,
-      );
-      const previewText = truncatePromptText(message.content);
-
-      return {
-        id: message.id,
-        index,
-        scrollTop,
-        previewText,
-      };
-    }).filter(Boolean);
-
-    if (baseItems.length === 0) {
-      setRailItems([]);
-      setActivePromptId(null);
-      return;
-    }
-
-    let activeIndex = 0;
-    let nextActiveId = baseItems[0]?.id || null;
-    for (const [itemIndex, item] of baseItems.entries()) {
-      if (item.scrollTop <= activeLine) {
-        nextActiveId = item.id;
-        activeIndex = itemIndex;
-      }
-    }
-
-    const maxVisibleCount = Math.max(Math.floor(railHeight / promptGap), 1);
-    const visibleCount = Math.min(baseItems.length, maxVisibleCount);
-    const visibleCenter = (visibleCount - 1) / 2;
-    const maxStartIndex = Math.max(baseItems.length - visibleCount, 0);
-    const startIndex = Math.min(
-      Math.max(Math.round(activeIndex - visibleCenter), 0),
-      maxStartIndex,
-    );
-    const visibleItems = baseItems.slice(startIndex, startIndex + visibleCount);
-
-    const nextItems = visibleItems.map((item, visibleIndex) => ({
-      ...item,
-      top: railCenter + (visibleIndex - visibleCenter) * promptGap,
-    }));
-
-    setRailItems(nextItems);
-    setActivePromptId(nextActiveId);
-  }, [activeChatId, promptMessages, streamRef, visible]);
-
-  const scheduleMeasure = useCallback(() => {
-    if (frameRef.current) return;
-    frameRef.current = requestAnimationFrame(measureRail);
-  }, [measureRail]);
-
-  useEffect(() => {
-    scheduleMeasure();
-  }, [messages, scheduleMeasure]);
-
-  useEffect(() => {
-    const scroller = streamRef.current;
-    const rail = railRef.current;
-    if (!visible || !scroller || !rail) return undefined;
-
-    const observer = new ResizeObserver(scheduleMeasure);
-    observer.observe(scroller);
-    observer.observe(rail);
-    if (scroller.firstElementChild) {
-      observer.observe(scroller.firstElementChild);
-    }
-
-    scroller.addEventListener("scroll", scheduleMeasure, { passive: true });
-    window.addEventListener("resize", scheduleMeasure);
-
-    return () => {
-      observer.disconnect();
-      scroller.removeEventListener("scroll", scheduleMeasure);
-      window.removeEventListener("resize", scheduleMeasure);
-    };
-  }, [scheduleMeasure, streamRef, visible]);
-
-  useEffect(
-    () => () => {
-      if (frameRef.current) cancelAnimationFrame(frameRef.current);
-    },
-    [],
-  );
-
-  if (!visible || promptMessages.length === 0) return null;
-
-  function jumpToPrompt(item) {
-    const scroller = streamRef.current;
-    if (!scroller) return;
-
-    const messageNode = Array.from(
-      scroller.querySelectorAll('[data-message-role="user"]'),
-    ).find((node) => node.dataset.messageId === String(item.id));
-    if (!messageNode) return;
-
-    const scrollerTop = scroller.getBoundingClientRect().top;
-    const messageTop = messageNode.getBoundingClientRect().top;
-    const maxScroll = Math.max(scroller.scrollHeight - scroller.clientHeight, 0);
-    const scrollTop = Math.min(
-      Math.max(scroller.scrollTop + messageTop - scrollerTop - 24, 0),
-      maxScroll,
-    );
-    const reducedMotion = window.matchMedia(
-      "(prefers-reduced-motion: reduce)",
-    ).matches;
-    const nearbyPrompt = (
-      Math.abs(scrollTop - scroller.scrollTop) <= scroller.clientHeight * 2
-    );
-
-    scroller.scrollTo({
-      top: scrollTop,
-      behavior: !reducedMotion && nearbyPrompt ? "smooth" : "auto",
-    });
-  }
-
-  return (
-    <nav className="prompt-nav-rail-wrap" aria-label="Prompt navigation">
-      <div
-        ref={railRef}
-        className="prompt-nav-rail"
-        onMouseLeave={() => {
-          setPreviewPromptId(null);
-        }}
-      >
-        {railItems.map((item) => {
-          const active = item.id === activePromptId;
-          const previewing = item.id === previewPromptId;
-
-          return (
-            <button
-              key={item.id}
-              type="button"
-              aria-label={`Jump to prompt ${item.index + 1}: ${item.previewText}`}
-              onClick={() => jumpToPrompt(item)}
-              onMouseEnter={() => setPreviewPromptId(item.id)}
-              onMouseLeave={() => setPreviewPromptId(null)}
-              onFocus={() => setPreviewPromptId(item.id)}
-              onBlur={() => setPreviewPromptId(null)}
-              className={cx(
-                "prompt-nav-tick",
-                active && "is-active",
-                previewing && "is-previewing",
-              )}
-              style={{ top: `${item.top}px` }}
-            >
-              <span aria-hidden="true" className="prompt-nav-line" />
-              <span
-                className={cx(
-                  "prompt-nav-preview",
-                  previewing && "is-open",
-                )}
-              >
-                {item.previewText}
-              </span>
-            </button>
-          );
-        })}
-      </div>
-    </nav>
-  );
-}
-
-const AssistantStatusLine = memo(function AssistantStatusLine({
-  reasoning,
-  reasoningStreaming,
-  waiting,
-  durationMs,
-  searching = false,
-}) {
-  const [open, setOpen] = useState(false);
-
-  const hasReasoning = Boolean(reasoning);
-  if (!hasReasoning && !waiting && !searching) return null;
-
-  const label = searching
-    ? "Searching"
-    : !hasReasoning
-      ? "Working"
-      : reasoningStreaming || !durationMs
-        ? "Thinking"
-        : `Thought for ${formatThoughtDuration(durationMs)}`;
-
-  const shimmering = searching || reasoningStreaming || (!hasReasoning && waiting);
-
-  return (
-    <div className="mb-4 max-w-3xl">
-      <button
-        type="button"
-        onClick={hasReasoning ? () => setOpen((value) => !value) : undefined}
-        aria-expanded={hasReasoning ? open : undefined}
-        disabled={!hasReasoning}
-        className={cx(
-          "inline-flex min-h-7 items-center rounded-md py-0 pr-2 text-xs font-medium text-neutral-500 focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/30",
-          CONTROL_MOTION,
-          hasReasoning ? "hover:text-neutral-300" : "cursor-default",
-        )}
-      >
-        <span
-          aria-hidden={!hasReasoning}
-          className={cx(
-            "-ml-1 mr-1 inline-flex shrink-0 items-center justify-center overflow-hidden",
-            "transition-[width,opacity] duration-200 ease-out",
-            hasReasoning ? "w-[15px] opacity-100" : "w-0 opacity-0",
-          )}
-        >
-          <ChevronDown
-            size={15}
-            className={cx("transition-transform duration-150 ease-out", !open && "-rotate-90")}
-          />
-        </span>
-        <StatusLabel label={label} shimmering={shimmering} />
-      </button>
-      {hasReasoning && open && (
-        <div className="mt-3 ml-[3px] border-l border-white/10 pl-[11px] text-pretty text-sm leading-7 text-neutral-500">
-          <ThinkingContent>{reasoning}</ThinkingContent>
-        </div>
-      )}
-    </div>
-  );
-});
-
-const MessageItem = memo(function MessageItem({
-  message,
-  streaming,
-  smoothStreaming,
-  reasoningStreaming,
-  reasoningDurationMs,
-  searching,
-  onCopy,
-  onRegenerate,
-  onEditUserMessage,
-  onDeleteUserMessage,
-}) {
-  const isUser = message.role === "user";
-  const messageAttachments = message.attachments || [];
-  const messageSources = uncitedSources(message.content, message.sources);
-  const articleRef = useRef(null);
-  const bodyRef = useRef(null);
-  const settledRef = useRef(0);
-  const paceWords = Boolean(streaming && smoothStreaming);
-  const { text: visibleContent, revealing } = useStreamReveal(
-    bodyRef,
-    paceWords,
-    message.content || "",
-    settledRef,
-  );
-  const [editing, setEditing] = useState(false);
-  const [draft, setDraft] = useState(message.content || "");
-  const [saving, setSaving] = useState(false);
-
-  useEffect(() => {
-    if (!editing) setDraft(message.content || "");
-  }, [editing, message.content]);
-
-  function captureScrollAnchor() {
-    const article = articleRef.current;
-    const scroller = article?.closest("section");
-    if (!article || !scroller) return null;
-    return {
-      article,
-      scroller,
-      top: article.getBoundingClientRect().top,
-    };
-  }
-
-  function restoreScrollAnchor(anchor) {
-    if (!anchor) return;
-    requestAnimationFrame(() => {
-      requestAnimationFrame(() => {
-        if (!anchor.article.isConnected || !anchor.scroller.isConnected) return;
-        const nextTop = anchor.article.getBoundingClientRect().top;
-        anchor.scroller.scrollTop += nextTop - anchor.top;
-      });
-    });
-  }
-
-  function startEditing() {
-    const anchor = captureScrollAnchor();
-    setEditing(true);
-    restoreScrollAnchor(anchor);
-  }
-
-  function cancelEditing() {
-    const anchor = captureScrollAnchor();
-    setEditing(false);
-    setDraft(message.content || "");
-    restoreScrollAnchor(anchor);
-  }
-
-  async function saveEdit() {
-    const next = draft.trim();
-    if (!next || next === message.content) {
-      cancelEditing();
-      return;
-    }
-    const anchor = captureScrollAnchor();
-    setSaving(true);
-    try {
-      await onEditUserMessage(message, next);
-      setEditing(false);
-      restoreScrollAnchor(anchor);
-    } finally {
-      setSaving(false);
-    }
-  }
-
-  if (isUser) {
-    return (
-      <article
-        ref={articleRef}
-        data-message-id={message.id}
-        data-message-role={message.role}
-        className={cx("group flex", editing ? "justify-stretch" : "justify-end")}
-      >
-        <div
-          className={cx(
-            "flex flex-col gap-2",
-            editing
-              ? "w-full items-stretch"
-              : "max-w-[78%] items-end sm:max-w-[68%]",
-          )}
-        >
-          {!editing && messageAttachments.length > 0 && (
-            <AttachmentChips attachments={messageAttachments} compact />
-          )}
-          <div
-            className={cx(
-              editing
-                ? "prompt-edit-surface w-full rounded-[28px] px-5 pb-5 pt-4"
-                : "chat-user-prompt whitespace-pre-wrap rounded-[22px] bg-neutral-200 px-4 py-3 text-pretty text-sm leading-6 text-neutral-950",
-              !editing && !(message.content || "").trim() && "hidden",
-            )}
-          >
-            {editing ? (
-              <>
-                <textarea
-                  value={draft}
-                  rows={Math.min(8, Math.max(3, draft.split("\n").length))}
-                  onChange={(event) => setDraft(event.target.value)}
-                  data-1p-ignore="true"
-                  onKeyDown={(event) => {
-                    if (event.key === "Enter" && event.metaKey) {
-                      event.preventDefault();
-                      saveEdit();
-                    }
-                    if (event.key === "Escape") {
-                      cancelEditing();
-                    }
-                  }}
-                  className="block max-h-[260px] min-h-[108px] w-full resize-none bg-transparent text-base leading-7 text-neutral-50 outline-none placeholder:text-neutral-500 sm:text-[17px]"
-                />
-                <div className="mt-5 flex justify-end gap-3">
-                  <button
-                    type="button"
-                    onClick={cancelEditing}
-                    className={cx(
-                      "inline-flex min-h-11 items-center rounded-full px-5 text-sm font-medium text-neutral-100 shadow-[0_0_0_1px_rgba(255,255,255,0.13)] hover:bg-white/[0.055] hover:shadow-[0_0_0_1px_rgba(255,255,255,0.2)] focus:outline-none focus-visible:ring-2 focus-visible:ring-white/20",
-                      CONTROL_MOTION,
-                    )}
-                  >
-                    Cancel
-                  </button>
-                  <button
-                    type="button"
-                    disabled={saving || !draft.trim()}
-                    onClick={saveEdit}
-                    className={cx(
-                      "inline-flex min-h-11 items-center rounded-full bg-white px-6 text-sm font-semibold text-neutral-950 hover:bg-neutral-100 focus:outline-none focus-visible:ring-2 focus-visible:ring-white/45 disabled:cursor-not-allowed disabled:bg-white/35 disabled:text-neutral-700 disabled:active:scale-100",
-                      CONTROL_MOTION,
-                    )}
-                  >
-                    {saving ? "Sending" : "Send"}
-                  </button>
-                </div>
-              </>
-            ) : (
-              message.content
-            )}
-          </div>
-          {!editing && (
-            <div className={cx("flex justify-center gap-1 opacity-0 group-hover:opacity-100 group-focus-within:opacity-100", FADE_MOTION)}>
-              <AssistantActionButton
-                label="Copy prompt"
-                onClick={() => onCopy(message)}
-              >
-                <Copy size={15} />
-              </AssistantActionButton>
-              <AssistantActionButton
-                label="Edit prompt"
-                onClick={startEditing}
-              >
-                <Pencil size={15} />
-              </AssistantActionButton>
-              <AssistantActionButton
-                label="Delete prompt"
-                onClick={() => onDeleteUserMessage(message)}
-              >
-                <Trash2 size={15} />
-              </AssistantActionButton>
-            </div>
-          )}
-        </div>
-      </article>
-    );
-  }
-
-  return (
-    <article
-      data-message-id={message.id}
-      data-message-role={message.role}
-      className="group max-w-none"
-    >
-      <AssistantStatusLine
-        reasoning={message.reasoning}
-        reasoningStreaming={reasoningStreaming}
-        waiting={!message.content}
-        durationMs={reasoningDurationMs}
-        searching={searching}
-      />
-      <div ref={bodyRef} className="max-w-3xl text-[15px] leading-7 text-neutral-100">
-        {message.content ? (
-          <MarkdownContent streaming={revealing} settledRef={settledRef}>
-            {visibleContent}
-          </MarkdownContent>
-        ) : null}
-      </div>
-      {message.content && !streaming && !revealing && (
-        <SourcePills sources={messageSources} className="mt-4 max-w-3xl" />
-      )}
-      {message.content && !streaming && !revealing && (
-        <div className={cx("mt-3 flex max-w-3xl justify-start gap-1 opacity-0 group-hover:opacity-100 group-focus-within:opacity-100", FADE_MOTION)}>
-          <AssistantActionButton label="Copy" onClick={() => onCopy(message)}>
-            <Copy size={15} />
-          </AssistantActionButton>
-          <AssistantActionButton
-            label="Regenerate"
-            onClick={() => onRegenerate(message.id)}
-          >
-            <RefreshCw size={15} />
-          </AssistantActionButton>
-          <ResponseInfoButton message={message} />
-        </div>
-      )}
-    </article>
-  );
-});
-
-function EmptyChatState() {
-  return (
-    <div className="min-h-[100dvh]" aria-hidden="true" />
-  );
-}
-
-function TemporaryChatButton({ active, onClick }) {
-  return (
-    <button
-      type="button"
-      data-tour="temp-chat-button"
-      aria-label={active ? "Temporary chat on" : "Temporary chat off"}
-      title={active ? "Temporary chat on" : "Temporary chat off"}
-      aria-pressed={active}
-      onClick={onClick}
-      className={cx(
-        "pointer-events-auto inline-flex h-10 w-10 items-center justify-center rounded-full text-[17px] leading-none focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45",
-        CONTROL_MOTION,
-        active
-          ? "text-neutral-100"
-          : "text-neutral-400 hover:text-neutral-100",
-      )}
-    >
-      <span className="t-icon-swap temp-chat-toggle-icon" data-state={active ? "b" : "a"}>
-        <span className="t-icon" data-icon="a" aria-hidden="true">
-          <i className="fi fi-rr-ghost" />
-        </span>
-        <span className="t-icon" data-icon="b" aria-hidden="true">
-          <i className="fi fi-sr-ghost" />
-        </span>
-      </span>
-    </button>
-  );
-}
-
-function TemporaryChatMarker({ visible }) {
-  if (!visible) return null;
-
-  return (
-    <div
-      className="pointer-events-none absolute left-4 top-16 z-20 flex h-10 w-10 items-center justify-center text-[18px] leading-none text-neutral-100 sm:left-8 sm:top-4 lg:left-10"
-      aria-hidden="true"
-    >
-      <i className="fi fi-rs-ghost" />
-    </div>
-  );
-}
-
-function MessageList({
-  messages,
-  activeChatId,
-  streamingMessageId,
-  smoothStreaming,
-  reasoningStreamingMessageId,
-  reasoningDurations,
-  searchingMessageId,
-  streamRef,
-  onScroll,
-  onWheel,
-  onTouchStart,
-  onTouchMove,
-  onCopy,
-  onRegenerate,
-  onEditUserMessage,
-  onDeleteUserMessage,
-}) {
-  return (
-    <section
-      ref={streamRef}
-      onScroll={onScroll}
-      onWheel={onWheel}
-      onTouchStart={onTouchStart}
-      onTouchMove={onTouchMove}
-      className="min-h-0 overflow-y-auto overscroll-contain px-4 py-8 sm:px-8 lg:px-10"
-    >
-      <div className="mx-auto flex w-full max-w-4xl flex-col gap-8">
-        {!activeChatId && messages.length === 0 ? (
-          <EmptyChatState />
-        ) : (
-          messages.map((message) => (
-            <MessageItem
-              key={message.id}
-              message={message}
-              streaming={message.id === streamingMessageId}
-              smoothStreaming={smoothStreaming}
-              reasoningStreaming={message.id === reasoningStreamingMessageId}
-              reasoningDurationMs={reasoningDurations[message.id]}
-              searching={message.id === searchingMessageId}
-              onCopy={onCopy}
-              onRegenerate={onRegenerate}
-              onEditUserMessage={onEditUserMessage}
-              onDeleteUserMessage={onDeleteUserMessage}
-            />
-          ))
-        )}
-      </div>
-    </section>
-  );
-}
-
-function StoryRail({
-  stories,
-  chapters,
-  activeStoryId,
-  activeChapterId,
-  mobileOpen,
-  onCloseMobile,
-  collapsed,
-  onCollapse,
-  onGoHome,
-  onCreateChapter,
-  onSelectStory,
-  onSelectChapter,
-  onNewStory,
-  onImportStory,
-  onRenameStory,
-  onExportStory,
-  onRenameChapter,
-  onDeleteStory,
-  onDeleteChapter,
-  onToggleChapterContext,
-  previousChatMode,
-  onChatModeChange,
-  navigationLocked = false,
-}) {
-  const [railScrolling, setRailScrolling] = useState(false);
-  const [railScrolled, setRailScrolled] = useState(false);
-  const [railHasMoreBelow, setRailHasMoreBelow] = useState(false);
-  const [renameTarget, setRenameTarget] = useState(null);
-  const [renameDraft, setRenameDraft] = useState("");
-  const [searchOpen, setSearchOpen] = useState(false);
-  const [recentsOpen, setRecentsOpen] = useState(true);
-  const [importingStory, setImportingStory] = useState(false);
-  const renameInputRef = useRef(null);
-  const importInputRef = useRef(null);
-  const skipRenameCommitRef = useRef(false);
-  const railRef = useRef(null);
-  const railScrollTimeoutRef = useRef(null);
-
-  function updateRailEdges(element) {
-    const bottomOffset = element.scrollHeight - element.clientHeight - element.scrollTop;
-    setRailScrolled(element.scrollTop > 2);
-    setRailHasMoreBelow(bottomOffset > 2);
-  }
-
-  useEffect(() => {
-    const rail = railRef.current;
-    if (!rail) return;
-
-    const frameId = requestAnimationFrame(() => updateRailEdges(rail));
-    const resizeObserver = typeof ResizeObserver === "undefined"
-      ? null
-      : new ResizeObserver(() => updateRailEdges(rail));
-    resizeObserver?.observe(rail);
-
-    return () => {
-      cancelAnimationFrame(frameId);
-      resizeObserver?.disconnect();
-    };
-  }, [stories.length, chapters.length, activeStoryId, recentsOpen]);
-
-  useEffect(
-    () => () => {
-      if (railScrollTimeoutRef.current) {
-        window.clearTimeout(railScrollTimeoutRef.current);
-      }
-    },
-    [],
-  );
-
-  function handleRailScroll(event) {
-    updateRailEdges(event.currentTarget);
-    setRailScrolling(true);
-    window.clearTimeout(railScrollTimeoutRef.current);
-    railScrollTimeoutRef.current = window.setTimeout(
-      () => setRailScrolling(false),
-      650,
-    );
-  }
-
-  function startRename(entityType, item) {
-    skipRenameCommitRef.current = false;
-    setRenameTarget({ entityType, id: item.id });
-    setRenameDraft(item.title || "");
-  }
-
-  function cancelRename() {
-    skipRenameCommitRef.current = true;
-    setRenameTarget(null);
-    setRenameDraft("");
-  }
-
-  async function commitRename(entityType, item) {
-    if (skipRenameCommitRef.current) {
-      skipRenameCommitRef.current = false;
-      return;
-    }
-
-    const nextTitle = renameDraft.trim();
-    if (!nextTitle || nextTitle === item.title) {
-      setRenameTarget(null);
-      setRenameDraft("");
-      return;
-    }
-
-    try {
-      if (entityType === "story") {
-        await onRenameStory(item, nextTitle);
-      } else {
-        await onRenameChapter(item, nextTitle);
-      }
-      setRenameTarget(null);
-      setRenameDraft("");
-    } catch {
-      requestAnimationFrame(() => renameInputRef.current?.focus());
-    }
-  }
-
-  async function importStoryFile(event) {
-    const file = event.target.files?.[0];
-    event.target.value = "";
-    if (!file) return;
-
-    setImportingStory(true);
-    try {
-      const imported = await onImportStory(file);
-      if (imported) onCloseMobile();
-    } finally {
-      setImportingStory(false);
-    }
-  }
-
-  return (
-    <>
-      <div
-        className={cx(
-          "fixed inset-0 z-30 bg-black/55 opacity-0 backdrop-blur-sm transition-[opacity,backdrop-filter] duration-200 ease-out lg:hidden",
-          mobileOpen ? "pointer-events-auto opacity-100" : "pointer-events-none",
-        )}
-        onClick={onCloseMobile}
-      />
-      <aside
-        className={cx(
-          "chat-sidebar t-resize fixed inset-y-0 left-0 z-40 flex w-[292px] flex-col overflow-hidden border-r border-line bg-[#080808] lg:static lg:z-auto lg:translate-x-0",
-          collapsed
-            ? "lg:w-0 lg:-translate-x-3 lg:border-r-0 lg:border-transparent lg:opacity-0"
-            : "lg:w-[276px] lg:opacity-100",
-          mobileOpen ? "translate-x-0" : "-translate-x-full",
-        )}
-      >
-        <div
-          className={cx(
-            "chat-sidebar-content flex h-full w-[292px] flex-col p-4 lg:w-[276px]",
-            collapsed
-              ? "lg:-translate-x-8 lg:opacity-0"
-              : "lg:translate-x-0 lg:opacity-100",
-          )}
-        >
-          <div className="mb-4 flex items-center justify-between gap-2 pl-2">
-            <div className="flex min-w-0 items-baseline gap-1.5">
-              <span className="truncate text-[19px] font-bold tracking-[-0.015em] text-white">
-                RouterChat
-              </span>
-              <span className="shrink-0 text-[19px] font-bold tracking-[-0.015em] text-neutral-500">
-                {APP_VERSION}
-              </span>
-            </div>
-
-            <div className="flex shrink-0 items-center">
-              <button
-                type="button"
-                aria-label="Search stories"
-                title="Search stories"
-                onClick={() => setSearchOpen(true)}
-                className={cx(
-                  "hidden h-10 w-10 shrink-0 items-center justify-center rounded-lg text-neutral-400 hover:bg-white/[0.08] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50 lg:inline-flex",
-                  CONTROL_MOTION,
-                )}
-              >
-                <MaskIcon src="/icons/search.png" size={19} />
-              </button>
-              <button
-                type="button"
-                aria-label="Collapse sidebar"
-                title="Collapse sidebar"
-                onClick={onCollapse}
-                className={cx(
-                  "hidden h-10 w-10 shrink-0 items-center justify-center rounded-lg text-neutral-400 hover:bg-white/[0.08] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50 lg:inline-flex",
-                  CONTROL_MOTION,
-                )}
-              >
-                <MaskIcon src="/icons/sidebar.png" size={15.5} />
-              </button>
-              <button
-                type="button"
-                aria-label="Close stories"
-                title="Close stories"
-                onClick={onCloseMobile}
-                className={cx(
-                  "inline-flex h-10 w-9 shrink-0 items-center justify-center rounded-lg text-neutral-400 hover:text-white focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50 lg:hidden",
-                  CONTROL_MOTION,
-                )}
-              >
-                <X size={19} />
-              </button>
-            </div>
-          </div>
-
-          <div className="mb-3.5 flex justify-center">
-            <SlidingTabs
-              options={CHAT_MODES}
-              value="write"
-              fromValue={previousChatMode}
-              onChange={onChatModeChange}
-              getValue={(mode) => mode.value}
-              getLabel={(mode) => mode.label}
-              ariaLabel="Interaction mode"
-              className="sidebar-mode-tabs"
-            />
-          </div>
-
-          <div className="mb-4">
-            <button
-              type="button"
-              data-tour="write-home-button"
-              onClick={() => {
-                if (navigationLocked) return;
-                onGoHome();
-                onCloseMobile();
-              }}
-              disabled={navigationLocked}
-              className={cx(
-                "flex h-9 w-full items-center gap-3 rounded-xl bg-transparent px-2 text-[15px] font-medium text-white hover:bg-white/[0.06] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45 disabled:cursor-not-allowed disabled:opacity-55",
-                CONTROL_MOTION,
-              )}
-            >
-              <span
-                aria-hidden="true"
-                className="grid h-5 w-5 shrink-0 place-items-center text-[17px] leading-none text-neutral-200"
-              >
-                <i className="fi fi-rr-home" />
-              </span>
-              Home
-            </button>
-
-            <button
-              type="button"
-              onClick={() => {
-                if (navigationLocked) return;
-                onNewStory();
-                onCloseMobile();
-              }}
-              disabled={navigationLocked}
-              className={cx(
-                "flex h-9 w-full items-center gap-3 rounded-xl bg-transparent px-2 text-[15px] font-medium text-white hover:bg-white/[0.06] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45 disabled:cursor-not-allowed disabled:opacity-55",
-                CONTROL_MOTION,
-              )}
-            >
-              <MaskIcon src="/icons/newbook.png" size={20} className="text-neutral-200" />
-              New story
-            </button>
-
-            <button
-              type="button"
-              onClick={() => importInputRef.current?.click()}
-              disabled={navigationLocked || importingStory}
-              className={cx(
-                "flex h-9 w-full items-center gap-3 rounded-xl bg-transparent px-2 text-[15px] font-medium text-white hover:bg-white/[0.06] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/45 disabled:cursor-not-allowed disabled:opacity-55",
-                CONTROL_MOTION,
-              )}
-            >
-              <span className="grid h-5 w-5 shrink-0 place-items-center">
-                <MaskIcon src="/icons/file-import.png" size={17} className="text-neutral-200" />
-              </span>
-              {importingStory ? "Importing story" : "Import story"}
-            </button>
-            <input
-              ref={importInputRef}
-              type="file"
-              aria-label="Import story file"
-              accept="application/json,.json"
-              className="hidden"
-              onChange={importStoryFile}
-            />
-          </div>
-
-          <div className="relative min-h-0 flex-1">
-            <nav
-              ref={railRef}
-              onScroll={handleRailScroll}
-              className={cx(
-                "chat-rail-scrollbar h-full space-y-3 overflow-y-auto pr-1",
-                railScrolling && "is-scrolling",
-              )}
-            >
-              {stories.length === 0 ? (
-              <div className="px-3 py-8 text-pretty text-sm leading-6 text-neutral-500">
-                Your stories will appear here.
-              </div>
-            ) : (
-              <SidebarGroup
-                label="Recents"
-                open={recentsOpen}
-                onToggle={() => setRecentsOpen((current) => !current)}
-              >
-                {stories.map((story) => {
-                  const active = story.id === activeStoryId;
-                  const renamingStory =
-                    renameTarget?.entityType === "story" && renameTarget.id === story.id;
-                  return (
-                    <div key={story.id} className="space-y-1">
-                      <div
-                        data-tour={active ? "write-story-rail" : undefined}
-                        className={cx(
-                          "group grid grid-cols-[minmax(0,1fr)_auto] items-center gap-1 rounded-2xl border border-transparent px-2 py-1",
-                          active ? "bg-white/[0.08] shadow-[var(--shadow-border)]" : "hover:bg-white/[0.045]",
-                        )}
-                      >
-                        {renamingStory ? (
-                          <div className="min-h-8 min-w-0 rounded-xl px-1 py-0.5">
-                            <input
-                              ref={renameInputRef}
-                              autoFocus
-                              aria-label="Rename story"
-                              data-1p-ignore="true"
-                              value={renameDraft}
-                              onChange={(event) => setRenameDraft(event.target.value)}
-                              onBlur={() => commitRename("story", story)}
-                              onKeyDown={(event) => {
-                                if (event.key === "Enter") event.currentTarget.blur();
-                                if (event.key === "Escape") {
-                                  event.preventDefault();
-                                  cancelRename();
-                                }
-                              }}
-                              className="block h-7 w-full min-w-0 rounded-md bg-white/[0.06] px-1.5 text-sm font-medium text-neutral-100 outline-none shadow-[var(--shadow-border)]"
-                            />
-                          </div>
-                        ) : (
-                          <button
-                            type="button"
-                            onClick={() => {
-                              if (navigationLocked) return;
-                              onSelectStory(story.id);
-                              onCloseMobile();
-                            }}
-                            disabled={navigationLocked}
-                            className="min-w-0 rounded-xl px-1 py-1 text-left focus:outline-none focus-visible:ring-2 focus-visible:ring-white/15"
-                          >
-                            <div className="truncate text-sm font-medium leading-4 text-neutral-100">
-                              {story.title}
-                            </div>
-                          </button>
-                        )}
-                        <StoryHistoryActions
-                          story={story}
-                          onRename={(item) => startRename("story", item)}
-                          onExport={onExportStory}
-                          onDelete={onDeleteStory}
-                          exportDisabled={navigationLocked}
-                        />
-                      </div>
-
-                      {active && (
-                        <div className="ml-3 space-y-1 border-l border-white/10 pl-2">
-                          <button
-                            type="button"
-                            data-tour="write-new-chapter-button"
-                            onClick={onCreateChapter}
-                            disabled={navigationLocked}
-                            className="mb-1 flex h-8 w-full items-center justify-center rounded-xl text-xs font-medium text-neutral-400 hover:bg-white/[0.045] hover:text-neutral-100 focus:outline-none focus-visible:ring-2 focus-visible:ring-white/15"
-                          >
-                            New chapter
-                          </button>
-                          {chapters.length === 0 ? (
-                            <div className="px-2 py-3 text-xs leading-5 text-neutral-600">
-                              No chapters yet.
-                            </div>
-                          ) : (
-                            chapters.map((chapter) => {
-                              const renamingChapter =
-                                renameTarget?.entityType === "chapter"
-                                && renameTarget.id === chapter.id;
-                              return (
-                                <div
-                                  key={chapter.id}
-                                  className={cx(
-                                    "group grid grid-cols-[minmax(0,1fr)_auto] items-center gap-1 rounded-xl px-2 py-1",
-                                    chapter.id === activeChapterId
-                                      ? chapter.disabled
-                                        ? "bg-white/[0.075] text-neutral-400"
-                                        : "bg-white/[0.075] text-neutral-100"
-                                      : chapter.disabled
-                                        ? "text-neutral-600 hover:bg-white/[0.04] hover:text-neutral-300"
-                                        : "text-neutral-400 hover:bg-white/[0.04] hover:text-neutral-100",
-                                  )}
-                                >
-                                  {renamingChapter ? (
-                                    <input
-                                      ref={renameInputRef}
-                                      autoFocus
-                                      aria-label="Rename chapter"
-                                      data-1p-ignore="true"
-                                      value={renameDraft}
-                                      onChange={(event) => setRenameDraft(event.target.value)}
-                                      onBlur={() => commitRename("chapter", chapter)}
-                                      onKeyDown={(event) => {
-                                        if (event.key === "Enter") event.currentTarget.blur();
-                                        if (event.key === "Escape") {
-                                          event.preventDefault();
-                                          cancelRename();
-                                        }
-                                      }}
-                                      className="block h-6 w-full min-w-0 rounded-md bg-white/[0.06] px-1.5 text-xs font-medium text-neutral-100 outline-none shadow-[var(--shadow-border)]"
-                                    />
-                                  ) : (
-                                    <button
-                                      type="button"
-                                      onClick={() => {
-                                        if (navigationLocked) return;
-                                        onSelectChapter(chapter.id);
-                                        onCloseMobile();
-                                      }}
-                                      disabled={navigationLocked}
-                                      className="flex min-w-0 items-center gap-1.5 text-left text-xs leading-5 focus:outline-none"
-                                    >
-                                      <span className="min-w-0 flex-1 truncate">
-                                        {chapter.title}
-                                      </span>
-                                      {chapter.disabled && (
-                                        <EyeOff
-                                          size={13}
-                                          className="shrink-0 text-neutral-500"
-                                          aria-hidden="true"
-                                        />
-                                      )}
-                                    </button>
-                                  )}
-                                  <ChapterHistoryActions
-                                    chapter={chapter}
-                                    onRename={(item) => startRename("chapter", item)}
-                                    onDelete={onDeleteChapter}
-                                    onToggleContext={onToggleChapterContext}
-                                  />
-                                </div>
-                              );
-                            })
-                          )}
-                        </div>
-                      )}
-                    </div>
-                  );
-                })}
-              </SidebarGroup>
-              )}
-            </nav>
-            <div
-              aria-hidden="true"
-              className={cx(
-                "sidebar-list-fade pointer-events-none absolute inset-x-0 top-0 h-6 bg-gradient-to-b from-[#080808]/95 to-transparent transition-opacity duration-150 ease-out",
-                railScrolled ? "opacity-100" : "opacity-0",
-              )}
-            />
-            <div
-              aria-hidden="true"
-              className={cx(
-                "sidebar-list-fade pointer-events-none absolute inset-x-0 bottom-0 h-7 bg-gradient-to-t from-[#080808]/95 to-transparent transition-opacity duration-150 ease-out",
-                railHasMoreBelow ? "opacity-100" : "opacity-0",
-              )}
-            />
-          </div>
-
-          <footer className="mt-1 -mb-2">
-            <FeedbackLink />
-          </footer>
-        </div>
-      </aside>
-
-      <SidebarSearchModal
-        open={searchOpen}
-        items={stories}
-        label="Search stories"
-        emptyText="Your stories will appear here."
-        noMatchText="No stories match that search."
-        fallbackTitle="Untitled story"
-        onSelect={(storyId) => {
-          setSearchOpen(false);
-          if (navigationLocked) return;
-          onSelectStory(storyId);
-          onCloseMobile();
-        }}
-        onClose={() => setSearchOpen(false)}
-      />
-    </>
-  );
-}
 
 function App() {
   const localAppSettings = readLocalAppSettings();
